+++ lib/multi_image_picker_view.dart
@@ -2,7 +2,10 @@ library multi_image_picker_view;
 
 export './src/multi_image_picker_view.dart';
 export './src/multi_image_picker_controller.dart';
-export './src/multi_image_picker_components.dart';
+export './src/widgets/multi_image_picker_initial_widget.dart';
+export './src/widgets/multi_image_picker_close_button.dart';
+export './src/widgets/multi_image_picker_image_widget.dart';
+export './src/widgets/multi_image_picker_add_more_button.dart';
 export './src/image_file.dart';
 export './src/image_file_view/image_file_view.dart';
 export './src/file_picker_extension.dart';
