+++ lib/src/image_file_view/io_preview.dart
@@ -1,3 +1,4 @@
+import 'dart:io';
 import 'package:flutter/material.dart';
 import '../image_file.dart';
 
@@ -11,8 +12,8 @@ class ImageFileView extends StatelessWidget {
   Widget build(BuildContext context) {
     return ClipRRect(
       borderRadius: BorderRadius.circular(4),
-      child: Image.network(
-        file.path!,
+      child: Image.file(
+        File(file.path!),
         fit: fit,
         errorBuilder: (context, error, stackTrace) {
           return const Center(child: Text('No Preview'));
