+++ lib/src/list_image_item.dart
@@ -4,56 +4,32 @@ import 'image_file.dart';
 import 'image_file_view/image_file_view.dart';
 
 class PreviewItem extends StatelessWidget {
-  const PreviewItem(
-      {super.key,
-      required this.file,
-      required this.fit,
-      this.boxDecoration,
-      // this.closeButtonBoxDecoration,
-      // required this.onDelete,
-      // this.closeButtonIcon,
-      // required this.showCloseButton,
-      // required this.closeButtonAlignment,
-      // required this.closeButtonMargin,
-      // required this.closeButtonPadding,
-      required this.defaultImageBorderRadius, this.closeButtonWidget});
+  const PreviewItem({
+    super.key,
+    required this.file,
+    required this.fit,
+    this.boxDecoration,
+    required this.defaultImageBorderRadius,
+  });
 
   final ImageFile file;
   final BoxFit fit;
   final BoxDecoration? boxDecoration;
-  // final BoxDecoration? closeButtonBoxDecoration;
-  // final EdgeInsetsGeometry closeButtonMargin;
-  // final EdgeInsetsGeometry closeButtonPadding;
-  // final Widget? closeButtonIcon;
-  // final bool showCloseButton;
+
   final BorderRadius? defaultImageBorderRadius;
-  // final Alignment closeButtonAlignment;
-  // final Function(ImageFile path) onDelete;
-  final Widget? closeButtonWidget;
 
   @override
   Widget build(BuildContext context) {
-    return Stack(
+    return Container(
       clipBehavior: Clip.antiAlias,
-      children: [
-        Positioned.fill(
-          child: Container(
-            clipBehavior: Clip.antiAlias,
-            decoration: boxDecoration ??
-                BoxDecoration(
-                  borderRadius: defaultImageBorderRadius,
-                ),
-            child: ImageFileView(
-              fit: fit,
-              file: file,
-            ),
+      decoration: boxDecoration ??
+          BoxDecoration(
+            borderRadius: defaultImageBorderRadius,
           ),
-        ),
-        if (closeButtonWidget != null)
-          Positioned.fill(
-            child: closeButtonWidget!,
-          ),
-      ],
+      child: ImageFileView(
+        fit: fit,
+        file: file,
+      ),
     );
   }
 }
